[Renderer] Don’t clamp text dimensions to strut
This was previously done to ensure e.g. the cursor height is always
the maximum character height, which prevents the cursor from resizing
as the user types, because that’s jank.

The problem with that is that it also messes with position calculations
and effectively introduces extra padding around text; to fix that, we
no longer clamp text dimensions; instead, any places in the codebase
that care about a font’s strut must now factor it into their calculations
explicitly.